requirements.txt
# Core AI & ML
google-generativeai>=0.3.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Computer Vision & Graphics
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
plotly>=5.17.0
scikit-image>=0.21.0

# UI & Interaction
ipywidgets>=8.1.0
ipython>=8.12.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
pyyaml>=6.0.0
jsonschema>=4.19.0
